renamed some structs
