+++ notan/crates/notan_app/src/builder.rs
@@ -1,7 +1,7 @@
 use crate::assets::{AssetManager, Loader};
 use crate::config::*;
 // use crate::graphics::Graphics;
-use crate::gfx::Graphics;
+use crate::graphics::Graphics;
 use crate::handlers::{
     AppCallback, AppHandler, DrawCallback, DrawHandler, EventCallback, EventHandler, SetupCallback,
 };
