+++ notan/crates/notan_app/src/empty.rs
@@ -1,6 +1,6 @@
 use crate::config::WindowConfig;
 use crate::graphics::prelude::*;
-use crate::graphics::{Graphics, GraphicsBackend, RenderTexture};
+use crate::graphics::{Device, DeviceBackend, RenderTexture};
 use crate::{App, Backend, BackendSystem, EventIterator, InitializeFn, WindowBackend};
 
 #[derive(Default)]
@@ -65,17 +65,17 @@ impl BackendSystem for EmptyBackend {
         }))
     }
 
-    fn get_graphics_backend(&self) -> Box<GraphicsBackend> {
-        Box::new(EmptyGraphicsBackend::default())
+    fn get_graphics_backend(&self) -> Box<DeviceBackend> {
+        Box::new(EmptyDeviceBackend::default())
     }
 }
 
 #[derive(Default)]
-struct EmptyGraphicsBackend {
+struct EmptyDeviceBackend {
     id_count: i32,
 }
 
-impl GraphicsBackend for EmptyGraphicsBackend {
+impl DeviceBackend for EmptyDeviceBackend {
     fn api_name(&self) -> &str {
         ""
     }
