+++ notan/crates/notan_graphics/src/prelude.rs
@@ -1,7 +1,7 @@
 pub use crate::buffer::*;
 pub use crate::color::*;
 pub use crate::commands::*;
-pub use crate::graphics::*;
+pub use crate::device::*;
 pub use crate::pipeline::*;
 pub use crate::render_texture::*;
 pub use crate::renderer::*;
